Revert "cdc_acm : Restrict writing more than 4 bytes into TX USB Endpoint."
This reverts commit f8c3f1717e15b53379ce9acb5acf2b96c536f819 [formerly 1da0a9eebd8b1a15e8247e0fb360171599d981e5].

The workaround caused a severe performance penalty, and only worked
for USB packets of 4-15 bytes in length (16+ byte packets weren't
subject to the hardware bug).  Single-byte packets (very common for
cdc_acm serial port transfers) would still be duplicated sometimes.

The upcoming DMA implementation does not share the performance
penalty, and also is not subject to the bug for those sizes of packets
(though it DOES still have a problem with single-byte packets!).

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>

Former-commit-id: 4a8a24fe39cd1437e00d845ea2573c0cd89e5e2d