arm: nxp: k6x: Add default partition table.
Recent changes (69255043, 91f67a13, 84628e8b, fa4a3932) add support
for a partition table in the flash.  Add support for this to the nxp
k6x dtsi file.  By default, code will occupy the entire flash.  By
setting a chosen node in an application, the code can be linked into
one of the partitions.  For example, and app could create a
'frdm_k64f.overlay' file at the top of their project with:

    / {
        chosen {
            zephyr,code-partition = &slot0_partition;
        };
    };

to place an application in slot 0.

Signed-off-by: David Brown <david.brown@linaro.org>

Former-commit-id: 795f068a2e28d4c4975f22e49522efe2916cc664