net: tcp: Add FIN timer when doing active close
The commit 2d7ac25bed53 [formerly 210c30805b9e] ("net: context: Close connection fast
if TIME_WAIT support is off") was not a proper way of closing
the connection. So if Zephyr closes the connection (active close),
then send FIN and install a timer that makes sure that if the peer
FIN + ACK is lost, we close the connection properly after a timeout.

Signed-off-by: Jukka Rissanen <jukka.rissanen@linux.intel.com>

Former-commit-id: e663c5df516db90af95093c95755cacb6a50d567